refactor (Calendar): EventType 위치 변경
