+++ src/main/java/com/sopt/cherrish/domain/calendar/domain/model/EventType.java
@@ -1,0 +1,5 @@
+package com.sopt.cherrish.domain.calendar.domain.model;
+
+public enum EventType {
+	PROCEDURE
+}
