--- src/main/java/com/sopt/cherrish/domain/calendar/presentation/dto/EventType.java
@@ -1,5 +1,0 @@
-package com.sopt.cherrish.domain.calendar.presentation.dto;
-
-public enum EventType {
-	PROCEDURE
-}
